+++ src/plib/gnw/gnw.cc
@@ -1486,7 +1486,8 @@ static void* colorOpen(const char* path)
 // 0x4C4298
 static int colorRead(void* handle, void* buf, size_t count)
 {
-    return db_fread(buf, 1, count, reinterpret_cast<DB_FILE*>(handle));
+    /* db_fread returns size_t; caller expects int — make conversion explicit */
+    return static_cast<int>(db_fread(buf, 1, count, reinterpret_cast<DB_FILE*>(handle)));
 }
 
 // 0x4C42A0
